Fixed delivery method deleting

+++ protected/modules/store/controllers/admin/DeliveryController.php
@@ -89,7 +89,12 @@ class DeliveryController extends SAdminController {
 			if (!empty($model))
 			{
 				foreach($model as $m)
-					$m->delete();
+				{
+					if($m->countOrders() == 0)
+						$m->delete();
+					else
+						throw new CHttpException(409, Yii::t('OrdersModule.admin','Ошибка удаления способа доставки. Он используется заказами.'));
+				}
 			}
 
 			if (!Yii::app()->request->isAjaxRequest)
